Display the VR keyboard on the throttle hand if the joystick hand is busy.
When JoystickEmul == 3 and the dominant hand is pressing grip, then this is
equivalent to having the dominant hand grabbing the joystick. In that case, it
makes no sense to display the VR keyboard on that hand since you don't want to
release the joystick to place the keyboard. So let's just display it on the
throttle hand instead.