Refactor VR Keyb buffers code, load green circles texture.
In order to fix the displacement suffered by inertia on the intersection point,
I'm going to load a "green circles" texture and place it in 3D as it should be.
However, in order to do that, I need to refactor the code that creates
rectangular meshes for the VR keyboard, so that I can use it to display these
circles too.