Fix inertia when displaying green-dot intersection guides.
The green dots I've been using to show which AC control will be clicked do not
respond to inertia, and that can be confusing while flying. To fix this, I'm
creating a small mesh that displays the green dots and this mesh is sent through
the regular 3D pipeline so that it acquires inertia, just like the rest of the
meshes.