Fix throttle emulation, place VR keyboard at finger tip.
The throttle no longer gets "stuck" if adjusted by the keyboard. It may even
co-exist with a joystick throttle now.

The DisplayMessage functions are now global to help debug VR from anywhere.

The VR keyboard is now displayed at the tip of the finger to make it easier
to place.