Properly display the VR Keyboard and gloves in the hangar.
The VR gloves and keyboard are displayed in viewspace coords -- this is exactly
what we do while in the Gunner Turret. This ensures that these elements are
always displayed relative to the camera.

The reason this wasn't working is because some meshes are not transformed with
the rest of the cockpit while in the hangar (like the landing gear!), so the
gloves were displayed *somewhere else* depending on which texture was captured
first.

I think displaying everything in viewspace coords should also work for regular
flight...