Keep the targeted craft inside the CMD.
In some ships, like the YT-2000, the VR gloves affected the transform on the
targeted craft. Further, sometimes the targeted craft appears outside the CMD!
This can be fixed by clearing the MeshTransform field when capturing the
miniature, but I couldn't fix the HMD transform getting transferred to the
targeted ship. I think that one comes from the CockpitLook hook.