New method to render a new reticle in VR, using billboards.
The RenderVRDots() method can be modified to move the dot 200m ahead of us, but
centered at the cockpit origin. Since the dot is a billboard, it will always
face the camera and since it's so far away, but still lies on the ship's forward
axis, it converges to the reticle. So, we can use this method to replace the
reticle without distortion and we can also use this to compute the 2D position
of the reticle even when it's not visible on the screen. This is important
because there's some hacky code that captures the reticle to figure out the
direction of the ship, and now we may no longer need to do that.

We could also fix the pips using this method, but more work is needed to project
the laser hardpoints and then there's the question of how to replace the other
reticle indicators. Another possibility is fixing the brackets around targeted
objects so that they don't rotate anymore with the camera (again, more
investigation is needed here). Still, baby steps, right?