Fix bracket rendering in VR mode.
The brackets are currently very "wiggly" in VR. This commit attempts to fix that
by making the brackets billboards that are always trying to be aligned with the
cockpit's up direction. This isn't a perfect solution, but the brackets are way
more stable now -- and the fix works in External view as well.