Removed SHADELESS_MAT and EMISSION_MAT. Fixed VR shaders.
Turns out that VR geometry was causing a white halo if the normal's alpha is set
to one (!) This was totally unexpected, but the fix was to set alpha to 0.5
just as in the hook_normals hook.

The HUD is still being shaded and maybe there's a few more artifacts in various
places.