Fix PixelShaderAnim/shadeless, code cleanup, change clear values for RTV.
Fixed the shadeless textures (i.e. cockpit damage on the glass) in
PixelShaderAnim.

Made the shadeless + transparent test in XwaD3dPixelShader more robust by
enabling the bIsTransparent constant. This is used for the cockpit glass.

Changed the clear values for the normBuf and depthBuf RTVs. The normBuf is now
initialized to 0 and depthBuf is now initialized to infinity.

Minor code cleanup in various areas.