Send the background to its own buffer, clear offscreenBuffer and blend.
In order to remove more halos, I'm moving the background to its own buffer and
the offscreenBuffer is then cleared so that it has alpha = 0. Then, this buffer
is shaded and the result is blended with the background. This removes most of
the remaining halos, but it's not a perfect solution.

For some reason, though, offscreenBuffer doesn't have an alpha == 0 when the
DeferredPass() (it's closer to 0.1 or so) is about to be called. I wonder where
it's getting this weird value from.