Render the reticle to its own RTV so that it can be blended cleanly.
The reticle was already being captured in VR mode. Now we're also capturing it
in non-VR mode so that it can be blended without artifacts on the deferred pass.

This change only affects pancake mode -- VR mode should continue to render the
reticle on its own 3D plane set at infinity.