Fix Missiles.
Missiles are complicated. They are rendered in two phases: the OPT/solid and a
shadeless/transparent layer, and one of those is also rendered as renderType = 2
(i.e. like a laser). Missiles are now handled separately on the shader and the
shadeless/transparent layer is now being redirected to transp layer #1. I've
removed dead blocks of code that are never hit because the OPT layer is renderer
in EffectsRenderer and the shadeless/transparent layer is rendered in
Direct3DDevice.

Also removed more dead references to right-eye buffers which are most likely not
used anymore -- anywhere.