Fix noGlassPS: initialize AspectRatio, Offset. Render DS2 Core Exp on transp1
The DS2 mission looked a little weird because some of the tubes were using the
AlphaIsntGlass property which activates the noGlassPS. This shader uses a float2
Offset that should've been initialized but wasn't... Who knows. Let's just make
sure it's initialized.

Also: the DS2 explosion is now rendered on transparent layer #1.