Fix bloom and shadeless properties. Fix DS2
The DS2 tunnel should now look correct (it was using the NoGlass shader which
now uses lightmaps properly), but the entrance and some tubes are rendered to
transpBuffer1.

The bloom was way too much for certain transparent textures, had to tone it
down.

The shadeless property wasn't actually getting applied at all!