Apply thresholds only on non-transparent lightmaps
I'm going ping-pong here. The fix for the banded holograms is not applying a
threshold based on value, but there are several OPTs that rely on that threshold
for bloom and without it, they appear shadeless.

This patch only applies the lightness-based threshold to lightmaps with
transparency, that appears to fix most cases. There's also a small fix that
allows textures to be tagged as transparent even if they don't have a material
file.