Don't use ForceShaded. Replace RTV pointer with an RTV enum to prevent crashes.
Looks like we don't need the ForceShaded property after all, we can just use
the smooth threshold in XwaD3dPixelShader for glass (the one that switches
between opaque and transparent areas) on transparent surfaces in general, like
external cockpits. We only have to separate lasers/explosions/etc on layer 1 and
cockpit transparency on layer 2. This way, we don't need to edit any mat files.
This also fixes the DS2 entrance tubes. Isn't this nice? Er... Go me? If only I
hadn't gotten myself into this mess in the first place. So: Go me! Aye?

Also: it's not a good idea to directly take the address of a ComPtr<>, there's
some random crashes. So let's instead use an RTV enum selector. This seems to
work.