Partial fix for VR Raytraced shadows.
By using the player's mobileObject's transform matrix right after the TLAS is
built, we can avoid the problems with the shadows -- except when inertia is on.

When inertia is enabled, it affects the lights because inertia is applied to the
cockpit, which is used to get the ship's orientation, but that also means that
lights are affected by inertia and when they illuminate external objects, the
shadows also inherit inertia.

We should not apply inertia to the light sources!