Fix artifacts that happen during a Hyperjump by clearing SSAOMask and SSMask
For some reason, both the SSAOMask and SSMask buffers have some transparency in
places when the hyperjump is activated and that causes artifacts to appear. I
added code to clear these RTVs in Direct3Device::BeginScene(), but I also added
the same code to clear them again when the background is captured. This appears
to have solved the problem.