Fix halos when jumping into hyperspace, minor refactor.
Some halos were still visible during the hyperspace sequence. To fix, this, I've
moved the hyper effects to the background buffer and then the blending is done
with alpha^2 even in shadeless areas. I made the shadeless test more correct
too. These changes made the SelectOffscreenBuffer() code much simpler, since we
no longer need to render the cockpit to a separate layer anymore.

The HyperComposeCommon shader is no longer blending the cockpit on top of the
hyper effects, since this is done already in the post-proc pass.

Removed some RTVs that were no longer used anywhere.