Selectively render to transp1RTV in the forward pass. Fix normal alpha on glass.
This is an alternative way to render glass in areas that have varying degrees of
alpha, like the CRS. I think this works, and it's correct: first shade the solid
areas, then put glass on top, shadeless.

Also fixed the normal's alpha in this section. It should be 0.5, not 1.0,
because hook_normals!