Fix missing lines in map mode.
In map mode there's no background, but we still merge it during the final
deferred shading pass. So what we do in this case is make sure that the
offscreenBuffer isn't wiped out when the background is captured, then we copy
the (resolved) offscreenBuffer onto the (resolved) background, so that in the
final deferred pass, we still have something to blend on top of.

Not very elegant, but it works. A better solution requires me putting some
effort into _really_ figuring out how the map mode works... but maybe some day.