Skip the draw calls that render the default starfield in VR mode.
The default starfield may still be visible in missions where the backdrops have
transparency (TFTC 1.3.4, B0M5, for instance). In VR, the default starfield is
messed up, so for now, it's better to skip those draw calls.

Notice how the detection of the default starfield is done by checking if we're
between the beginning of the frame, and the first call to Direct3D::Execute()