Disable the Sky Cylinder in a couple of places.
If VR is ON and we're parked in the hangar and the cockpit camera is active,
then we don't need to render the Sky Cylinder (we don't need it for cockpit
cameras in general)

If VR is OFF and we're parked in the hangar we don't need this at all. (We don't
need it at all for Non-VR, but having it available helps debug things).