Fix Concourse crash when the HD Concourse is enabled (WIP).
The HD Concourse is rendered using the D2D1 interface to make an RTV directly
from _offscreenBuffer -- but in VR, this buffer is an array. So we need a new
buffer (offscreenBufferHd) that isn't an array and is also the proper size
(2560x1440). This fixed the crash, but there's more...

Further, OffscreenSurface and FrontbufferSurface return a DC (from GetDC()) that
is used to render the HD Concourse. This DC needs to use the proper resolution
(2560x1400 again) or else the resulting image will be cropped.

There's still a bug where sometimes the HD images are not erased between frames
and a trail is created. That's why this fix is still WIP.