Fix non-VR mode display, fix trails in VR mode.
_offscreenBufferHdBackground must be copied to _offscreenBufferHd when
appropriate, and we need to clear _offscreenBufferHd too.

When playing in non-VR mode, we still need to use the old width/height in
OffscreenSurface and FrontbufferSurface, or the display will not be stretched
to cover the whole window.

The Tech Room is now broken, but there's no more trails in the HD Concourse.