Render the Default Starfield for VR and non-VR modes using ExternalHUD
Instead of rendering the reticle, we can use ExternalHUD to render the skybox.
I'm using code from the HyperStreak shader to convert the UV coords into a valid
direction in space that can be used to sample the SkyBox. The direction vector
was modified to match the direction vector used in PixelShaderVRGeom. I'm also
copying the Heading and ViewMatrix from RenderSkyBox(): Why work twice if I had
already solved this problem?