Fix the "background swim effect" for the default starfield.
The center of the screen does not correspond to the center of motion in this
game: it's displaced vertically depending on the current ship. I've known this
for quite a while, and I've been using a "y_center" factor to compensate for
that. However, if the external camera is enabled, this factor was not used
because I thought it wasn't necessary. Well, it turns out that this was causing
the "background swim effect" in VR. Which means that both the cockpit camera and
exterior views need the same y_center factor or we'll see the background move
in weird ways.

Now it's time to fix this problem once and for all... everywhere.