Fix the "swimming background" for external view in VR.
So, it turns out that for VR, we need to set y_center to 0 when the external
camera is set, and use y_center for the cockpit camera; otherwise the background
will "swim". Which means that in external view, the center of motion is at the
center of the screen. This contrasts with the previous fix where I had to use
y_center all the time. But in the previous case, I'm using 3D direction vectors
computed for a post-proc "screen", whereas here it's actually a 3D object being
placed at infinity.

Funny thing is that in external camera, distant objects have way too much
parallax. This suggests that XWA may be scalling down objects (?) Or maybe we're
looking at "miniatures" in this mode (?) Either way, this bug has been fixed.
Finally!