Do not apply the Heading to the Default Starfield when Alt+U is pressed.
Pressing Alt+U when a target is selected switches the camera to the target in
external view. In this case, the motion of the player's ship should not be
applied to the default starfield. I've added some code to detect this situation
and hopefully fix this without breaking other stuff.