Fix the Default Starfield orientation for the Gunner Turret
The Gunner Turret transform chain for the speed effect *substracts* the heading
of the ship, making the matrix relative to the direction of the ship. This is
fine for the speed effect, because we want the relative motion wrt to the ship's
heading for the speed particles. But for a global frame of reference, like the
Default Starfield, we need to keep the ship's heading in the transform chain.