Fix the black bar that at the bottom of the screen in exterior view/hangar.
In the hangar view, the first light is used to cast cockpit shadows using shadow
mapping. But if the external view is displayed, then there's no cockpit and we
get an ugly black bar near the bottom. We need to invalidate the first light in
this situation to fix this artifact.