Capture the external camera state to fix default background artifacts.
The default background disappears if no HUD is displayed and there's no visible
backdrops. In this case, Direct3Device::Execute() is not called, which in turn
means the default background is never rendered. To fix this, I added a flag that
tells us if the default background has been rendered in the current frame. If it
hasn't been rendered, then we render it during PrimarySurface::Flip().

Further, the background rendered in this way is static. We also need to capture
the state of the external camera while rendering 3D content to provide the right
values to the default background.

This is not an elegant solution, but I can't think of a better option at the
moment.