Fix the blending mode for the default starfield.
We're now rendering the default starfield at the end of the frame using data
captured at the beginning. Since the starfield may be rendering over an already
populated backgroundBuffer, we need to preserve the contents of said buffer
and only "fill in the gaps". We achieve this by enabling blending and setting
the blending mode for the source color as INV_DEST_ALPHA; meaning that only
areas in the destination that are 0 will enable the source color.