Call UpdateViewMatrix() in Direct3DDevice::BeginScene()
It some situations, Direct3DDevice::Execute() is not called at all when
rendering a frame (there's no backdrops in the current mission and no HUD
either, for instance). If this happens, we may lose VR tracking! To prevent this
we need to call UpdateViewMatrix() in Direct3DDevice::BeginScene() which is
always called (why didn't I do this before?)