Fix transparent DC elements.
Blending must be enabled for transparent DC elements (obviously!)

Something not so obvious: transparent DC elements should not write to the depth
buffer, and they should be "passthrough". This is important because the cockpit
sometimes gets rendered before the background and it must write normal = 0. When
this happens, that area becomes shadeless and displays the backdrops. That's why
we were seeing the backdrops through the T/B's cockpit. By avoding the depth
buffer write, we allow distant objects to overwrite the normal and fix the
situation.