Fix the Speed Effect v2 for VR.
In VR the input are 2D pancake coordinates. These coords are back-projected into
metric 3D and then projected back into 2D using the SteamVR projection matrices
for each eye. These last reprojected 2D coordinates need to be used in the Pixel
Shader to properly sample the depth buffer in VR.

Also, the viewport needs to be the same used during the forward render, but that
was already fine -- I just did a minor cleanup of the code.