Allow textures with the word "Cockpit" in them.
Some OPTs can now have textures with the word "Cockpit" in them. This confused
the system because it was using that same word to detect cockpits -- even on
Exterior OPTs, for instance.

The fix here is to make sure that the texture is referencing a Cockpit OPT and
for good measure, we should be doing a caseless comparison too.