Remove unnecessary/unbound samplers.
The DirectX debug layer shows a bunch of warnings because I used many many
samplers that I never actually bound to anything. It's not a catastrophe,
because all unbound samplers default to the NULL sampler and it works alright,
but it's better to get rid of these warnings so that we can see more warnings
and fix them as well.