Prevent Release() underflow by avoiding deletion of null-named textures.
Some textures have a name of size 0. I have no idea what that means, but if
these textures are deleted, then we get a resource underflow error (i.e. some
SRVs are released even after they have already a refcount of 0). So, for now,
I'm just avoiding the deletion to prevent the underflow. Not a real fix, but
maybe later we'll figure it out.