Fix Hyperzoom when exiting hyperspace.
The Hyperzoom effect was semi-broken. Since the stellar background is now
rendered on its own buffer, it was missing from the hyperzoom effect. The fix
here is to send the 3D background content *and* the stellar background to the
Hyperzoom shader so that they can be blended before applying the zoom effect.