Fix interaction between semi=transparent backdrops and DefaultStarfield.dds
Before the DefaulStarfield.dds was added, the game used XWA's default starfield
which was mostly black witha few dots. When a semi-transparent backdrop was
rendered on top of that, these backdrops looked fine.

Now, DefaulStarfield.dds is rendered *after* all the backdrops have been
rendered. And it will end up covering semi-transparent backdrops. To prevent
this, we need to discard black areas of DefaulStarfield.dds and it needs to be
similar to XWA's original starfield (mostly black, a few stars).

Not a great solution, but it works.