Single-char fix solves Release() underflow random crashes.
The following line:

	auto context = _deviceResources->_d3dDeviceContext;

Was creating a local variable that was a ComPtr object. When the variable goes
out of scope, it gets destroyed and it calls Release() automatically, decreasing
the refcount of _d3dDeviceContext too when it was still being used everywhere.
At some point, when _d3dDeviceContext is actually released (at the end of a
mission, for instance), the system attempts a Release() on an object that
already has 0 refcount and Hilarity Ensues.

It may also be possible that _d3dDeviceContext may have been randomly destroyed
by the system since, technically, it had a refcount of 0. Again, Hilarity
Ensues.

By adding an &, we're making a reference to the original object, which does not
get destroyed when leaving the current scope.

Problem solved. By a single char.

I feel so stupid now.

This bug is ancient too. It was added by commit 8cf116af on Mar 1, 2022.

Well, at least it was fixed. Better late than never, eh?