Fix the map in VR mode
The map is completely broken in VR. Part of the reason is that 2D and 3D content
is interleaved and there's no hint as to the depth of the 2D contents (map
icons, grid), so there's no way to place these elements properly in VR.

Instead, let's render the map as if we were in pancake mode and let
DrawHUDVertices() display it flat. Not great, but at least it will be useable.