Prevent multiple Resolves() when rendering the VR map.
The VR map renders to the offscreen DC buffer, but we need to make sure it's
only resolved once or its contents will get clobbered. To do this, I made sure
that the relevant buffers are resolved and cleared during DrawHUDVertices().

Also, the brackets are now visible in the VR map too. I just had to force them
to render using the regular non-VR path and made them all go to the DC buffer.