Fix blending artifacts between the Skybox and the Background.
Backgrounds are rendered before the skybox. To prevent replaicng the existing
background with the skybox, I was using a discard for pixels with low lightness.
This was causing a hard edge that looked like "pepper noise" in some areas. The
fix is to avoid hard edges by using a custom blending operation that combines
the alpha from the skybox and the inverse alpha from the existing background.

The VR path becomes more complicated: it needs two passes. One to render the
Skybox and another one to blend it with the current background.