Don't use Alt+2 for VR controller fire. Pull down on VR thr stops ship.
Using Alt+2 to activate firing with the VR controller causes some issues (other
synthetic keystrokes are blocked while firing). Instead, let's actually push the
trigger button on the virtual joystick to fire. This allows us to use the VR
throttle to simultaneously modify the speed.

Also, let's map "push down" on the VR throttle to something. In this case, let's
do a Backslash to stop the ship.