Add disable_cockpit_renders to functionally disable DC.
When dynamic_cockpit_enabled is set to 0, a number of artifacts happen. I can't
just stop using the DC buffers anymore because the reticle uses them to fix some
transparency issues. Instead, the plan is to always have "DC on", but stop
replacing the cockpit textures. That's what disable_cockpit_renders does, and
it's just the first step (eventually, I'm just going to remove
g_bDynCockpitEnabled).