Remove g_bDynCockpitEnabled (it's always enabled now).
DC has become an integral part of the Effects system, to the point where if it's
disable, several artifacts happen. We can still functionally "disable" it by not
replacing textures inside the cockpit, but we need to keep the rest of the
system working.

Also, disable_cockpit_renders is now properly read from the settings (the
previous commit was reading the wrong string).