Rename control var to g_bEnableEnhancedHUD, display enhanced HUD text (WIP).
g_bEnableEnhancedBrackets is now g_bEnableEnhancedHUD to make the naming
consistent with the settings.

I also added some code to display the word "Target" at the center of the current
target. I added a cache of strings, but I don't think that's actually necessasy.
This is WIP and it's mostly for testing how new text can be displayed. Also,
this won't work for VR.