Add s_ExeCraftTable and properly compute and display shields.
I was already using s_ExeCraftTable for the POVOffsets, but I was computing the
offsets into the table manually. With this table, it should be much easier to
access the definitions for all the ships in the game. This table can be used
to read the max shield value, thus making it possible to compute the shield
percentage.

This code was translated from code that was graciously provided by Jeremy.