Display the current subcomponent's name by parsing the OPT tree.
When OPTs are loaded, we can also parse their tree to find the MeshDescriptor
entry. This entry has a uint16_t* that has the index into Strings.txt which
has the name of the component. By creating a map from mesh index to Strings.txt
index per object species, we can look up the name of the current subcomponent
bracket.