Limit the smallest square around which text can be rendered. Make cargo empty.
For far away ships, we want to use a square that is bigger than the bracket so
that the text is readable. I've used 50 pixels here, but this probably needs to
be adjusted to account for resolution, etc.

If a ship has no cargo, then we shouldn't display "Unknown" before it's
inspected (this happens with T/Fs, for instance).