Display enhanced text in VR using an extra bracket (WIP).
The extra bracket is added when the current bracket is the target. The enhanced
text is rendered on fixed positions in the enhanced HUD buffer. The enhanced
HUD buffer is now square and is VR_ENHANCED_HUD_BUFFER_SIZE pixels in size.

The text is not readable for distant objects, so this isn't a great solution,
but the text is displayed nonetheless.