Fix the smoke. Finally.
Smoke is rendered with the same texture as the engine glow, which is bad because
I made the engine glows bloom for a long time. This means that smoke *also*
blooms and that's just plain wrong.

To solve this, I've hooked the function that XWA uses to render the engine
glows. XWA calls this function in two places and both have been hooked and
redirected. The hook simply sets a flag to signal that engine glows are going to
be rendered on the next draw() call. The draw() call sets this flag back to
false, always. This appears to work!

Thanks to Jeremy for the information on the XWA render engine glow function.