Use Dynamic Cockpit regions to render the enhanced HUD in VR.
This code uses TARGETED_OBJ_NAME_SRC_IDX to render the targeted craft's name
on the enhanced VR bracket. I also had to modify the source like this:

	source_def = 256,128, 72,10, 184,25

Otherwise some names appear cropped.

There's some dead code here and some refactoring needed to display more regions,
but at least this is working fine now. The text is also displayed with the right
ratio and is upright as long as it's not near the edges of the visual field.