Properly display targeted ship's name on the bracket edge.
To place the targeted ship's name on the top edge of the bracket, we need to
compute a rotation angle around the X axis. This angle is given by a rect
triangle created by halfWidthOPT and the desired depth (65km). Also fixed the
shader interpolation to place the text at the center of the bracket.

In order for this to work properly, we need the following too:

; Slot 33: CMD targeted object's name
source_def = 256,128, 72,15, 184,28

In Dynamic_Cockpit_Internal_Areas.cfg