Display text at the top edge of the current target bracket.
I decided to copy the transform chain we use in RenderVRBrackets(). That way,
the text follows the bracket. The text is always displayed at the same distance,
so that way it won't change it's size. To find the position where the text
should be placed, we displace a fake vertex from the Dot Buffer to the edge of
the quad. Then the fake vertex is processed through the regular
RenderVRBrackets() transform chain to find the place where the text should be
placed. Still needs some work, but the text shows alright, finally.