Fix RenderVRHUD(), add separate flags for reticle and enhanced HUD.
The VR HUD was no longer visible if enhanced brackets were enabled because they
used the same flag to avoid redundant renders.

Also moved the VR reticle to infinity (65km away) so that it actually appears
well-centered on the reticle. It might still need to compensate for headset roll,
though.