Make brackets compatible with the Gunner Turret.
I just copied the transform chain in RenderVRBrackets(). In order for this to
work, I need to remove the headset view matrix and replace the world view
transform with an identity matrix. That way, brackets just keep their location,
which is already in the view coord sys. Finally, I need to add swapScale to make
the coordinates compatible with the DirectX pipeline.