Parse g_xwa_text and extract DC data for the enhanced HUD
g_xwa_text holds all the text that will be rendered, and we know the locations
of the DC source boxes. Thus, by intersecting the chars in g_xwa_text with the
boxes we can extract computer-readable strings that can be used for the enhanced
HUD. We could also fix DC text boxes later, using the same approach.

Also removed some dead code from the VR path for the enhanced HUD.