Re-arrange text, render overshields bar, fix name color for TFTC.
All the text is now displayed in a single column, top center.

When shields are above 100%, an "overshields" bar is rendered on top.

In TFTC, sometimes ships don't have names! In that case, the algorithm for
finding the color of the name was broken. This has now been fixed and the color
in that case is automatically enhanced (brightness++, saturation++) to keep
readability.

The following optional settings were added to SSAO.cfg:

	enhanced_hud_overshds_col = 0.49, 0.97, 1
	enhanced_hud_enhance_name_col = 1

Setting enhanced_hud_enhance_name_col to 0 keeps the original (faded) color for
the names of ships in TFTC (when no ship name is present).