Render the Enhanced HUD text to its own buffer.
Based on this commit: 4982dbba35d9a76feef047b72949cd65ceb2b54b

If we render the enhanced HUD to its own buffer, then we can do layering and
have the cockpit block the text (instead of always displaying it -- remember,
don't let anything get in the way of my cockpit!). This will also help when
porting the enhanced HUD to VR (because we can use that buffer as an SRV to the
shader that displays this information in VR).