Render transparency for text, add optional dark background on text.
Turns out the text was losing its alpha channel because of the way the font was
setting its AntiAlias mode (ClearType). Setting it to grayscale preserves the
anti-aliasing, and the alpha channel.

With that out of the way, the code can be massaged so that it optionally renders
a dark rectangle behind the text to make it more readable. This feature is
disabled by default. To enable it, add:

	enhanced_hud_enable_bg_text_box = 1

To SSAO.cfg. Finally, did some minor refactor and changed how text is enhanced.