Parse shields front, back and rename target stats.
We can now parse the shields front/back for the player, but I had to rename the
existing target stats to avoid confusion. Also, the dynamic cockpit internal
area for the front shields needs to be shrunk vertically or the "S" for the
shield recharge rate sometimes gets caught as well.