Telemetry: Parse player shields from DC areas that match xwaText chars.
We can parse the contents of the xwaText vector when they intersect specific DC
areas. This way, we can get the values of the shields as currently displayed and
use them for Telemetry.

For some reason, some of these chars are repeated. So I added an algorithm to
remove duplicates.

Also, the following DC area should be modified as follows:

	; Slot 25: Front shields strength
	source_def = 128,128,  34,13, 63,25