Add g_DCSubRegions to break DC regions that contain multiple values.
DC "Sub Regions" are portions of DC regions that contain multiple values. For
instance the speed-and-throttle DC region can be broken in the throttle and
speed component. These sub regions are initialized when the parent DC region is
captured in Direct3DDevice.cpp and the sub-regions aren't exact. Instead, text
chars that overlap with the subregions are aggregated to rebuild the original
data in ExtractDCText().

Another thing that is worth noting is that the engine sometimes renders the same
text multiple times (sometimes with a small horizontal 1-pixel offset). So we
also need to keep track of chars and coords to reject duplicates. Fun stuff...
Pretty, pretty, preeetty fun...