Add AUTOSIZE_SPEED_SRC DC element. Correctly compute font height.
Added the AUTOSIZE_SPEED_SRC Dynamic Cockpit element. In order for this to work,
I needed to compute the _correct_ font height. This is done in two steps: first
the screen-height is computed, then the in-game height. Then we can use this
height to properly capture the speed. Tested this with a couple of fonts and
couple of different in-game resolutions.

The font heights are recomputed every time OnSizeChanged() is called.