There's no missiles/chaff counters in the PPG. Let's null those boxes.
When the PPG is playing, instead of having the regular digits for missiles and
countermeasures, we get a string telling us the number of laps. If we're in the
PPG, these AUTOSIZE boxes need to be set to 0.

Also: Jeremy just told me how to check if we're in a PPG mission.