Add AUTOSIZE_MISSILES_MIS_SRC DC source region.
This source box captures all missiles -- including primary and secondary. I also
had to move the CHAFF source box a little to the right, otherwise the last
digits of the secondary missiles were sometimes captured.

This box is primarily for the Missile Boat, but it can also capture "raw"
missile data for all other craft (by "raw", I mean it will capture the colon and
spaces as well, not just the digits).